Found Deletion Bug: Merges are not local
For some reason, I thought that you only need to delete/merge nodes
only locally for deletion and it doesn't need to e.g., go all the way
to the root. The example in this commit, delete_with_merge2(), shows
that this is in fact not possible. We have to restructure the code
like with insertion.